adds Esri module and abstract class
