+++ server/app/services/geo/esri/EsriClient.java
@@ -8,98 +8,31 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableList;
 import com.jayway.jsonpath.JsonPath;
 import com.jayway.jsonpath.ReadContext;
-import com.typesafe.config.Config;
 import java.time.Clock;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
-import java.util.concurrent.atomic.AtomicInteger;
-import javax.inject.Inject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import play.libs.Json;
-import play.libs.ws.WSBodyReadables;
-import play.libs.ws.WSBodyWritables;
-import play.libs.ws.WSClient;
-import play.libs.ws.WSRequest;
-import play.libs.ws.WSResponse;
 import services.Address;
-import services.AddressField;
 import services.geo.AddressLocation;
 import services.geo.AddressSuggestion;
 import services.geo.AddressSuggestionGroup;
 import services.geo.ServiceAreaInclusion;
 import services.geo.ServiceAreaState;
 
-/**
- * Provides methods for handling reqeusts to external Esri geo and map layer services for getting
- * address suggestions and service area validation for a given address.
- *
- * <p>@see <a
- * href="https://developers.arcgis.com/rest/geocode/api-reference/geocoding-find-address-candidates.htm">Find
- * Address Candidates</a>
- *
- * @see <a
- *     href="https://developers.arcgis.com/rest/services-reference/enterprise/query-feature-service-layer-.htm">Query
- *     (Map Service/Layer)</a>
- */
-public class EsriClient implements WSBodyReadables, WSBodyWritables {
-  private final Clock clock;
-  private final EsriServiceAreaValidationConfig esriServiceAreaValidationConfig;
-  private final WSClient ws;
-
-  private static final String ESRI_CONTENT_TYPE = "application/json";
-  // Specify output fields to return in the geocoding response with the outFields parameter
-  private static final String ESRI_FIND_ADDRESS_CANDIDATES_OUT_FIELDS =
-      "Address, SubAddr, City, RegionAbbr, Postal";
-  // The service supports responses in JSON or PJSON format. You can specify the response format
-  // using the f parameter. This is a required parameter
-  private static final String ESRI_RESPONSE_FORMAT = "json";
-  @VisibleForTesting Optional<String> ESRI_FIND_ADDRESS_CANDIDATES_URL;
-  private int ESRI_EXTERNAL_CALL_TRIES;
+/** An abstract class for working with external Esri services. */
+public abstract class EsriClient {
+  final Clock clock;
+  final EsriServiceAreaValidationConfig esriServiceAreaValidationConfig;
 
   private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
-  @Inject
-  public EsriClient(
-      Config configuration,
-      Clock clock,
-      EsriServiceAreaValidationConfig esriServiceAreaValidationConfig,
-      WSClient ws) {
+  public EsriClient(Clock clock, EsriServiceAreaValidationConfig esriServiceAreaValidationConfig) {
     this.clock = checkNotNull(clock);
-    this.ws = checkNotNull(ws);
-    this.ESRI_FIND_ADDRESS_CANDIDATES_URL =
-        configuration.hasPath("esri_find_address_candidates_url")
-            ? Optional.of(configuration.getString("esri_find_address_candidates_url"))
-            : Optional.empty();
-    this.ESRI_EXTERNAL_CALL_TRIES =
-        configuration.hasPath("esri_external_call_tries")
-            ? configuration.getInt("esri_external_call_tries")
-            : 3;
     this.esriServiceAreaValidationConfig = checkNotNull(esriServiceAreaValidationConfig);
-  }
-
-  /** Retries failed requests up to the provided value */
-  private CompletionStage<WSResponse> tryRequest(WSRequest request, int retries) {
-    CompletionStage<WSResponse> responsePromise = request.get();
-    AtomicInteger retryCount = new AtomicInteger(retries);
-    responsePromise.handle(
-        (result, error) -> {
-          if (error != null || result.getStatus() != 200) {
-            logger.error(
-                "Esri API error: {}", error != null ? error.toString() : result.getStatusText());
-            if (retries > 0) {
-              return tryRequest(request, retryCount.decrementAndGet());
-            } else {
-              return responsePromise;
-            }
-          } else {
-            return CompletableFuture.completedFuture(result);
-          }
-        });
-
-    return responsePromise;
   }
 
   /**
@@ -121,45 +54,7 @@ public class EsriClient implements WSBodyReadables, WSBodyWritables {
    * to the configured ESRI_EXTERNAL_CALL_TRIES. If ESRI_EXTERNAL_CALL_TRIES is not configured, the
    * tries default to 3.
    */
-  @VisibleForTesting
-  CompletionStage<Optional<JsonNode>> fetchAddressSuggestions(ObjectNode addressJson) {
-    if (this.ESRI_FIND_ADDRESS_CANDIDATES_URL.isEmpty()) {
-      return CompletableFuture.completedFuture(Optional.empty());
-    }
-    WSRequest request = ws.url(this.ESRI_FIND_ADDRESS_CANDIDATES_URL.get());
-    request.setContentType(ESRI_CONTENT_TYPE);
-    request.addQueryParameter("outFields", ESRI_FIND_ADDRESS_CANDIDATES_OUT_FIELDS);
-    // "f" stands for "format", options are json and pjson (PrettyJson)
-    request.addQueryParameter("f", ESRI_RESPONSE_FORMAT);
-    // limit max locations to 3 to keep the size down, since CF stores the suggestions in the user
-    // session
-    request.addQueryParameter("maxLocations", "3");
-    Optional<String> address =
-        Optional.ofNullable(addressJson.findPath(AddressField.STREET.getValue()).textValue());
-    Optional<String> address2 =
-        Optional.ofNullable(addressJson.findPath(AddressField.LINE2.getValue()).textValue());
-    Optional<String> city =
-        Optional.ofNullable(addressJson.findPath(AddressField.CITY.getValue()).textValue());
-    Optional<String> region =
-        Optional.ofNullable(addressJson.findPath(AddressField.STATE.getValue()).textValue());
-    Optional<String> postal =
-        Optional.ofNullable(addressJson.findPath(AddressField.ZIP.getValue()).textValue());
-    address.ifPresent(val -> request.addQueryParameter("address", val));
-    address2.ifPresent(val -> request.addQueryParameter("address2", val));
-    city.ifPresent(val -> request.addQueryParameter("city", val));
-    region.ifPresent(val -> request.addQueryParameter("region", val));
-    postal.ifPresent(val -> request.addQueryParameter("postal", val));
-
-    return tryRequest(request, this.ESRI_EXTERNAL_CALL_TRIES)
-        .thenApply(
-            res -> {
-              // return empty if still failing after retries
-              if (res.getStatus() != 200) {
-                return Optional.empty();
-              }
-              return Optional.of(res.getBody(json()));
-            });
-  }
+  abstract CompletionStage<Optional<JsonNode>> fetchAddressSuggestions(ObjectNode addressJson);
 
   /**
    * Calls an external Esri service to get address suggestions given the provided {@link Address}.
@@ -180,9 +75,9 @@ public class EsriClient implements WSBodyReadables, WSBodyWritables {
         .thenApply(
             (maybeJson) -> {
               if (maybeJson.isEmpty()) {
-                logger.error(
-                    "EsriClient.fetchAddressSuggestions JSON response is empty. Called by"
-                        + " EsriClient.getAddressSuggestions. Address = {}",
+                this.logger.error(
+                    "RealEsriClient.fetchAddressSuggestions JSON response is empty. Called by"
+                        + " RealEsriClient.getAddressSuggestions. Address = {}",
                     address);
                 return Optional.empty();
               }
@@ -247,35 +142,8 @@ public class EsriClient implements WSBodyReadables, WSBodyWritables {
    * tries default to 3.
    */
   @VisibleForTesting
-  CompletionStage<Optional<JsonNode>> fetchServiceAreaFeatures(
-      AddressLocation location, String validationUrl) {
-    WSRequest request = ws.url(validationUrl);
-    request.setContentType(ESRI_CONTENT_TYPE);
-    // "f" stands for "format", options are json and pjson (PrettyJson)
-    request.addQueryParameter("f", ESRI_RESPONSE_FORMAT);
-    request.addQueryParameter("geometryType", "esriGeometryPoint");
-    request.addQueryParameter("returnGeometry", "false");
-    request.addQueryParameter("outFields", "*");
-    // "inSR" specifies the spatial reference for the service to use
-    request.addQueryParameter("inSR", location.getWellKnownId().toString());
-    String geo = "{'x':";
-    geo += location.getLongitude();
-    geo += ",'y':";
-    geo += location.getLatitude();
-    geo += ",'spatialReference':";
-    geo += location.getWellKnownId();
-    request.addQueryParameter("geometry", geo);
-
-    return tryRequest(request, this.ESRI_EXTERNAL_CALL_TRIES)
-        .thenApply(
-            res -> {
-              // return empty if still failing after retries
-              if (res.getStatus() != 200) {
-                return Optional.empty();
-              }
-              return Optional.of(res.getBody(json()));
-            });
-  }
+  abstract CompletionStage<Optional<JsonNode>> fetchServiceAreaFeatures(
+      AddressLocation location, String validationUrl);
 
   /**
    * Calls an external Esri service to get the service areas of the provided {@link
@@ -290,7 +158,7 @@ public class EsriClient implements WSBodyReadables, WSBodyWritables {
 
     if (!esriServiceAreaValidationConfig.isConfigurationValid()) {
       logger.error(
-          "Error calling EsriClient.getServiceAreaInclusionGroups. Error:"
+          "Error calling RealEsriClient.getServiceAreaInclusionGroups. Error:"
               + " EsriServiceAreaValidationConfig.getImmutableMap() returned empty.");
       serviceAreaInclusionBuilder
           .setServiceAreaId(esriServiceAreaValidationOption.getId())
@@ -309,8 +177,8 @@ public class EsriClient implements WSBodyReadables, WSBodyWritables {
             (maybeJson) -> {
               if (maybeJson.isEmpty()) {
                 logger.error(
-                    "EsriClient.fetchServiceAreaFeatures JSON response is empty. Called by"
-                        + " EsriClient.isAddressLocationInServiceArea."
+                    "RealEsriClient.fetchServiceAreaFeatures JSON response is empty. Called by"
+                        + " RealEsriClient.isAddressLocationInServiceArea."
                         + " EsriServiceAreaValidationOption = {}. AddressLocation = {}",
                     esriServiceAreaValidationOption,
                     location);
