+++ server/conf/application.conf
@@ -105,6 +105,7 @@ play.modules {
   enabled += modules.DatabaseSeedModule
   enabled += modules.ProgramCreationModule
   enabled += modules.DurableJobModule
+  enabled += modules.EsriModule
 
   # If there are any built-in modules that you want to disable, you can list them here.
   #disabled += ""
