+++ server/test/services/geo/esri/EsriServiceAreaValidationOptionTest.java
@@ -1,20 +1,14 @@
 package services.geo.esri;
 
 import static org.junit.Assert.assertEquals;
-import static play.mvc.Results.ok;
 
 import com.google.common.collect.ImmutableList;
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
-import java.io.IOException;
 import java.time.Clock;
 import java.time.ZoneId;
-import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import play.libs.ws.WSClient;
-import play.routing.RoutingDsl;
-import play.server.Server;
 import services.geo.AddressLocation;
 import services.geo.ServiceAreaInclusion;
 
@@ -25,9 +19,7 @@ public class EsriServiceAreaValidationOptionTest {
   private AddressLocation location;
 
   // setup for service area validation
-  private Server server;
-  private WSClient ws;
-  private EsriClient client;
+  private FakeEsriClient client;
   private ImmutableList<ServiceAreaInclusion> inclusionGroup;
 
   @Before
@@ -47,36 +39,18 @@ public class EsriServiceAreaValidationOptionTest {
 
     location =
         AddressLocation.builder()
-            .setLongitude(-122.3360380354971)
-            .setLatitude(47.578374020558954)
+            .setLongitude(-100.0)
+            .setLatitude(100.0)
             .setWellKnownId(4326)
             .build();
 
-    server =
-        Server.forRouter(
-            (components) ->
-                RoutingDsl.fromComponents(components)
-                    .GET("/query")
-                    .routingTo(request -> ok().sendResource("esri/serviceAreaFeatures.json"))
-                    .build());
-    ws = play.test.WSTestClient.newClient(server.httpPort());
-
-    client = new EsriClient(config, clock, esriServiceAreaValidationConfig, ws);
+    client = new FakeEsriClient(clock, esriServiceAreaValidationConfig);
 
     inclusionGroup =
         client
             .getServiceAreaInclusionGroup(esriServiceAreaValidationOption, location)
             .toCompletableFuture()
             .join();
-  }
-
-  @After
-  public void tearDown() throws IOException {
-    try {
-      ws.close();
-    } finally {
-      server.stop();
-    }
   }
 
   @Test
